negative/requirements.txt
functions-framework==3.*
google-cloud-language==2.13.0
google-cloud-secret-manager==2.16.0
requests==2.31.0
gunicorn==20.1.0